requirements.txt
asgiref==3.2.10
astroid==2.3.3
autopep8==1.5
bleach==3.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
dj-database-url==0.5.0
Django==3.1.14
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-heroku==0.3.1
django-humanize==0.1.2
django-ipware==2.1.0
django-qr-code==4.2.0
docutils==0.16
future==0.18.2
geocoder==1.38.1
geoip2==3.0.0
gunicorn==20.0.4
humanize==2.0.0
idna==2.9
ip2geotools==0.1.5
IP2Location==8.4.1
isort==4.3.21
keyring==21.2.0
lazy-object-proxy==1.4.3
lxml==4.9.1
maxminddb==1.5.2
mccabe==0.6.1
packaging==20.3
Pillow==9.0.1
pip-review==1.0
pkginfo==1.5.0.1
pycodestyle==2.5.0
Pygments==2.7.4
pylint==2.4.4
pyparsing==2.4.6
pyquery==1.4.1
python-decouple==3.3
pytz==2019.3
pywin32-ctypes==0.2.0
qrcode==6.1
ratelim==0.1.6
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
tqdm==4.43.0
twine==3.1.1
typing==3.7.4.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.0.1
